# message: optimize lookupAndFormat function for better performance